feat: check for valid zip decorator
